Renomeando coisas e organizando códigos
