+++ src/controllers/index.js
@@ -4,6 +4,7 @@ export * from './addresses.js'
 export * from './controlPanel.js'
 export * from './hire.js'
 export * from './serviceCategory.js'
+export * from './servicePanel.js'
 export * from './services.js'
 export * from './user.js'
 export * from './userProfile.js'
