+++ src/controllers/user.js
@@ -22,117 +22,6 @@ const redirectIfAuthenticated = (request, response, next) => {
 @Router.Route('/user')
 export class User
 {
-	@Router.Get('/weighted')
-	static async weightedRating({ params }, response)
-	{
-		const sql = `SELECT
-			r.receiverId,
-			u.fullname,
-			AVG(r.rating) AS average,
-			COUNT(*) AS count,
-			1.0 * (:confidence * :alpha + SUM(r.rating)) / (:confidence + COUNT(*)) AS normalizedScore
-			FROM reviews r
-			INNER JOIN users u ON u.id = r.receiverId
-			GROUP BY receiverId`
-
-		try {
-			const ranking = await sequelize.query(sql, {
-				type: sequelize.QueryTypes.SELECT,
-				replacements: {
-					id: params.id,
-					// Represents a prior for the average of the stars.
-					alpha: 2.5,
-					// Represents how confident we're in our prior.
-					// It is equivalent to a number of observations.
-					confidence: 5.0
-				}
-			})
-
-			//return response.json(ranking)
-			return response.render('weightedRanking.pug', {
-				ranking
-			})
-		} catch (e) {
-			return response.status(400).render('error.pug', {
-				status: '0x05',
-				error: 'Erro ao computar relatório',
-				message: 'Erro desconhecido.',
-				stack: e.stack
-			})
-		}
-	}
-
-	@Router.Get('/:id/report/services')
-	static async serviceReport({ params }, response)
-	{
-		const sql = `SELECT users.fullname,
-			serviceCategories.name AS category,
-			services.title AS title,
-			SUM(contracts.totalPrice) AS sum,
-			COUNT(*) AS count,
-			AVG(contracts.totalPrice) AS avg
-			FROM users
-			INNER JOIN services ON users.id = services.userId
-			INNER JOIN contracts ON contracts.serviceId = services.id
-			INNER JOIN serviceCategories ON serviceCategories.id = services.serviceCategoryId
-			WHERE users.id = :id
-			GROUP BY services.id
-			ORDER BY SUM(contracts.totalPrice) DESC`
-
-		try {
-			const rank = await sequelize.query(sql, {
-				type: sequelize.QueryTypes.SELECT,
-				replacements: {
-					id: params.id
-				}
-			})
-			return response.json(rank)
-		} catch (e) {
-			return response.status(400).render('error.pug', {
-				status: '0x05',
-				error: 'Erro ao computar relatório',
-				message: 'Erro desconhecido.',
-				stack: e.stack
-			})
-		}
-	}
-
-	@Router.Get('/:id/report')
-	static async userReport({ params }, response)
-	{
-		const sql = `SELECT
-				users.fullname AS fullname,
-				serviceCategories.name AS category,
-				services.title AS title,
-				contracts.totalPrice AS price,
-				services.basePrice AS baseprice,
-				serviceCategories.pricingType AS pricingType
-			FROM users
-			INNER JOIN contracts ON "contracts.userId" = users.id
-			INNER JOIN services ON "contracts.serviceId" = services.id
-			INNER JOIN serviceCategories ON serviceCategories.id = "services.serviceCategoryId"
-			WHERE users.id = :id
-			ORDER BY contracts.totalPrice DESC`
-
-		try {
-			const rank = await sequelize.query(sql, {
-				type: sequelize.QueryTypes.SELECT,
-				replacements: {
-					id: params.id
-				}
-			})
-
-			return response.json(rank)
-		} catch (e) {
-			return response.status(400).render('error.pug', {
-				status: '0x05',
-				error: 'Erro ao computar relatório',
-				message: 'Erro desconhecido.',
-				stack: e.stack
-			})
-		}
-	}
-
 	@Router.Get('/logout')
 	@Router.Post('/logout')
 	static async logout(request, response)
@@ -143,13 +32,18 @@ export class User
 	}
 
 	@Router.Get('/login', [ redirectIfAuthenticated ])
+	static async loginPage (request, response)
+	{
+		return response.render('login.pug')
+	}
+
 	@Router.Post('/login', [ redirectIfAuthenticated ])
 	static async login(request, response)
 	{
 		const { email, password } = request.body
 
 		if (! email || ! password)
-			return response.render('login.pug')
+			return response.render('login.pug', { message: 'Preencha todos os campos.' })
 
 		return await $User.findOne({
 			where: { email },
@@ -168,19 +62,22 @@ export class User
 	}
 
 	@Router.Get('/register', [ redirectIfAuthenticated, upload.single('photoPath') ])
+	static async register (request, response)
+	{
+		return response.render('register.pug', { provinces })
+	}
+
 	@Router.Post('/register', [ redirectIfAuthenticated, upload.single('photoPath') ])
-	static async register({ body, file, session }, response)
+	static async register ({ body, file, session }, response)
 	{
 		if (_.isEmpty(body))
 			return response.status(400).render('register.pug', {
-				provinces: provinces,
-				error: ['Preencha todos os campos.']
+				provinces, error: [ 'Preencha todos os campos.' ]
 			})
 
 		if (body.password !== body.confirmPassword)
 			return response.status(400).render('register.pug', {
-				provinces: provinces,
-				error: [ 'As senhas inseridas devem ser iguais.' ]
+				provinces, error: [ 'As senhas inseridas devem ser iguais.' ]
 			})
 
 		await sequelize.transaction(async (transaction) => {
@@ -212,8 +109,7 @@ export class User
 					})
 				} catch (e) {
 					response.status(400).render('register.pug', {
-						provinces: provinces,
-						error: [ 'Erro ao cadastrar endereço ou cartão.' ]
+						provinces, error: [ 'Erro ao cadastrar endereço ou cartão.' ]
 					})
 				}
 			} catch (e) {
