--- src/views/controlPanel/index.pug
@@ -1,9 +1,0 @@
-extends ../index.pug
-
-block content
-	.card.w-100
-		.card-body
-			h1.card-title Painel administrativo
-			a(href='/cpanel/pending').card-link.btn.btn-primary Serviços pendentes
-			a(href='/cpanel/active').card-link.btn.btn-primary Serviços ativos
-			a(href='/cpanel/history').card-link.btn.btn-primary Histórico de serviços e avaliações
