+++ src/views/controlPanel/weightedRanking.pug
@@ -1,0 +1,26 @@
+extends ../index.pug
+
+block content
+	.card.w-100
+		.card-body
+			h4.card-title Ranking
+			table.table.table-hover
+				caption Relação de notas
+				thead
+					tr
+						th(scope='col') #
+						th(scope='col') Usuário
+						th(scope='col') Nota normalizada
+						th(scope='col') Nota média
+						th(scope='col') Avaliações
+				tbody
+					each user, index in ranking
+						tr
+							th(scope='row')= index + 1
+							td= user.fullname
+							td= user.normalizedScore.toFixed(2)
+							td= user.average.toFixed(2)
+							td= user.count
+					else
+						tr
+							td(colspan=5).center-text Não há avaliações cadastradas
