+++ src/views/servicePanel/active.pug
@@ -31,6 +31,6 @@ block inner
 							p
 								.fa.mr-2.fa-fw.fa-lg.fa-location-arrow
 								= `${ct.address.street}, ${ct.address.number}, ${ct.address.neighborhood}, ${ct.address.city}, ${ct.address.province}`
-						a(href=`/cpanel/active/complete/${ct.id}`).list-group-item.list-group-item-action
+						a(href=`/svpanel/active/complete/${ct.id}`).list-group-item.list-group-item-action
 							.fa.mr-2.fa-fw.fa-lg.fa-check
 							| Marcar como completo e avaliar
