+++ src/views/servicePanel/pending.pug
@@ -31,9 +31,9 @@ block inner
 							p
 								.fa.mr-2.fa-fw.fa-lg.fa-location-arrow
 								= `${ct.address.street}, ${ct.address.number}, ${ct.address.neighborhood}, ${ct.address.city}, ${ct.address.province}`
-						a(href=`/cpanel/pending/accept/${ct.id}`).list-group-item.list-group-item-action
+						a(href=`/svpanel/pending/accept/${ct.id}`).list-group-item.list-group-item-action
 							.fa.mr-2.fa-fw.fa-lg.fa-check
 							| Aceitar contrato
-						a(href=`/cpanel/pending/refuse/${ct.id}`).list-group-item.list-group-item-action
+						a(href=`/svpanel/pending/refuse/${ct.id}`).list-group-item.list-group-item-action
 							.fa.mr-2.fa-fw.fa-lg.fa-times
 							| Recusar contrato
