+++ src/views/topbar.pug
@@ -10,7 +10,7 @@ nav.navbar.navbar-expand.navbar-dark.bg-dark
 		.navbar-nav
 			+navlink('/services', 'Serviços')
 			if isLoggedIn
-				+navlink('/cpanel', 'Painel administrativo')
+				+navlink('/svpanel', 'Painel administrativo')
 				+navlink('/user/profile', 'Meu perfil')
 				+navlink('/user/logout', 'Encerrar sessão')
 			else
