--- src/views/weightedRanking.pug
@@ -1,24 +1,0 @@
-extends index.pug
-
-block content
-	div(style={ 'margin-bottom': '20px' }).card.w-100
-		.card-body
-			h4.card-title Ranking
-			table.table.table-stripped
-				caption Relação de notas
-				thead
-					tr
-						th(scope='col') #
-						th(scope='col') Usuário
-						th(scope='col') Nota normalizada
-						th(scope='col') Nota média
-						th(scope='col') Avaliações
-				tbody
-					each user, index in ranking
-						th(scope='row')= index + 1
-						td= user.fullname
-						td= user.normalizedScore.toFixed(2)
-						td= user.average.toFixed(2)
-						td= user.count
-					else
-						td(colspan=5).center-text Não há avaliações cadastradas
